iRfcb: Tools for Managing Imaging FlowCytobot (IFCB) Data

A comprehensive suite of tools for managing, processing, and analyzing data from the IFCB. I R FlowCytobot ('iRfcb') supports quality control, geospatial analysis, and preparation of IFCB data for publication in databases like https://www.gbif.org, https://www.obis.org, https://emodnet.ec.europa.eu/en, https://shark.smhi.se/en/, and https://www.ecotaxa.org. The package integrates with the MATLAB 'ifcb-analysis' tool, which is described in Sosik and Olson (2007) doi:10.4319/lom.2007.5.204 , and provides features for working with raw, manually classified, and machine learning–classified image datasets. Key functionalities include image extraction, particle size distribution analysis, taxonomic data handling, and biomass concentration calculations, essential for plankton research.

See also

Author

Maintainer: Anders Torstensson anders.torstensson@smhi.se (ORCID) (Swedish Meteorological and Hydrological Institute)

Other contributors:

Kendra Hayashi (ORCID) [contributor]

Jamie Enslein [contributor]

Raphael Kudela (ORCID) [contributor]

Alle Lie (ORCID) [contributor]

Jayme Smith (ORCID) [contributor]

DTO-BioFlow (Horizon Europe, HORIZON-MISS-2022-OCEAN-01-07) [funder]

SBDI (Swedish Research Council, 2019-00242) [funder]