Get Shark Column Names

This function reads SHARK column names from a specified tab-separated values (TSV) file included in the package. These columns are used for submitting IFCB data to https://sharkweb.smhi.se/.

Usage

ifcb_get_shark_colnames(minimal = FALSE)

Arguments

minimal

A logical value indicating whether to load only the minimal set of column names required for data submission to SHARK. Default is FALSE.

Value

An empty data frame containing the SHARK column names.

Details

For a detailed example of a data submission, see ifcb_get_shark_example.

See also

Examples

shark_colnames <- ifcb_get_shark_colnames()
print(shark_colnames)
#>  [1] MYEAR                     STATN                    
#>  [3] SAMPLING_PLATFORM         PROJ                     
#>  [5] ORDERER                   SHIPC                    
#>  [7] CRUISE_NO                 DATE_TIME                
#>  [9] SDATE                     STIME                    
#> [11] TIMEZONE                  LATIT                    
#> [13] LONGI                     POSYS                    
#> [15] WADEP                     MPROG                    
#> [17] MNDEP                     MXDEP                    
#> [19] SLABO                     ACKR_SMP                 
#> [21] SMTYP                     PDMET                    
#> [23] SMVOL                     METFP                    
#> [25] IFCBNO                    SMPNO                    
#> [27] LATNM                     SFLAG                    
#> [29] TRPHY                     APHIA_ID                 
#> [31] IMAGE_VERIFICATION        COUNT                    
#> [33] ABUND                     BIOVOL                   
#> [35] C_CONC                    QFLAG                    
#> [37] COEFF                     CLASS_NAME               
#> [39] CLASS_F1                  UNCLASSIFIED_COUNTS      
#> [41] UNCLASSIFIED_ABUNDANCE    UNCLASSIFIED_VOLUME      
#> [43] METOA                     COUNTPROG                
#> [45] ALABO                     ACKR_ANA                 
#> [47] ANADATE                   METDC                    
#> [49] TRAINING_SET              TRAINING_SET_ANNOTATED_BY
#> [51] CLASSIFIER_CREATED_BY     CLASSIFIER_USED          
#> [53] MANUAL_QC_DATE            PRE_FILTER_SIZE          
#> [55] PH_FB                     CHL_FB                   
#> [57] CDOM_FB                   PHYC_FB                  
#> [59] PHER_FB                   WATERFLOW_FB             
#> [61] TURB_FB                   PCO2_FB                  
#> [63] TEMP_FB                   PSAL_FB                  
#> [65] OSAT_FB                   DOXY_FB                  
#> <0 rows> (or 0-length row.names)

shark_colnames_minimal <- ifcb_get_shark_colnames(minimal = TRUE)
print(shark_colnames_minimal)
#>  [1] MYEAR               STATN               PROJ               
#>  [4] ORDERER             SHIPC               SDATE              
#>  [7] STIME               LATIT               LONGI              
#> [10] POSYS               MNDEP               MXDEP              
#> [13] SLABO               ACKR_SMP            SMTYP              
#> [16] SMVOL               IFCBNO              SMPNO              
#> [19] LATNM               SFLAG               TRPHY              
#> [22] IMAGE_VERIFICATION  COUNT               QFLAG              
#> [25] COEFF               CLASS_F1            UNCLASSIFIED_COUNTS
#> [28] METOA               COUNTPROG           TRAINING_SET       
#> [31] ALABO               ACKR_ANA            ANADATE            
#> [34] METDC               CLASSIFIER_USED    
#> <0 rows> (or 0-length row.names)