Get Shark Column Example

This function reads a SHARK submission example from a file included in the package. This format is used for submitting IFCB data to https://sharkweb.smhi.se/.

Usage

ifcb_get_shark_example()

Value

A data frame containing example data following the SHARK submission format.

See also

Examples

shark_example <- ifcb_get_shark_example()
print(shark_example)
#>   MYEAR                  STATN SAMPLING_PLATFORM              PROJ ORDERER
#> 1  2022 RV_FB_D20220713T175838              IFCB IFCB, DTO, JERICO    SMHI
#> 2  2022 RV_FB_D20220713T175838              IFCB IFCB, DTO, JERICO    SMHI
#> 3  2022 RV_FB_D20220713T175838              IFCB IFCB, DTO, JERICO    SMHI
#> 4  2022 RV_FB_D20220713T175838              IFCB IFCB, DTO, JERICO    SMHI
#> 5  2022 RV_FB_D20220713T175838            SveaFB IFCB, DTO, JERICO    SMHI
#>   SHIPC CRUISE_NO   DATE_TIME      SDATE    STIME TIMEZONE   LATIT LONGI POSYS
#> 1  77SE        12 2,02207E+13 2022-07-13 17:58:38      UTC 55.4826  12.7   GPS
#> 2  77SE        12 2,02207E+13 2022-07-13 17:58:38      UTC 55.4826  12.7   GPS
#> 3  77SE        12 2,02207E+13 2022-07-13 17:58:38      UTC 55.4826  12.7   GPS
#> 4  77SE        12 2,02207E+13 2022-07-13 17:58:38      UTC 55.4826  12.7   GPS
#> 5  77SE        12 2,02207E+13 2022-07-13 17:58:38      UTC 55.4826  12.7   GPS
#>   WADEP MPROG MNDEP MXDEP SLABO ACKR_SMP    SMTYP PDMET SMVOL METFP  IFCBNO
#> 1    NA  PROJ     4     4  SMHI        N FerryBox   MIX 4.875  None IFCB134
#> 2    NA  PROJ     4     4  SMHI        N FerryBox   MIX 4.875  None IFCB134
#> 3    NA  PROJ     4     4  SMHI        N FerryBox   MIX 4.875  None IFCB134
#> 4    NA  PROJ     4     4  SMHI        N FerryBox   MIX 4.875  None IFCB134
#> 5    NA  PROJ     4     4  SMHI        N FerryBox          NA  None        
#>              SMPNO                 LATNM SFLAG TRPHY APHIA_ID
#> 1 D20220713T175838 Prorocentrum cordatum          AU   232376
#> 2 D20220713T175838   Skeletonema marinoi          AU   376667
#> 3 D20220713T175838           Strombidium    CF    MX   101195
#> 4 D20220713T175838                                         NA
#> 5 D20220713T175838                                         NA
#>   IMAGE_VERIFICATION COUNT  ABUND      BIOVOL    C_CONC QFLAG COEFF
#> 1                 NA    18 3692.6 0.008255370 1.1148900    NA 205.1
#> 2                 NA     6 1230.9 0.000382593 0.0211181    NA 205.1
#> 3                 NA    16 3282.3 0.010800000 1.4152000    NA 205.1
#> 4                 NA    NA     NA          NA        NA    NA 205.1
#> 5                 NA    NA     NA          NA        NA    NA    NA
#>              CLASS_NAME CLASS_F1 UNCLASSIFIED_COUNTS UNCLASSIFIED_ABUNDANCE
#> 1 Prorocentrum_cordatum   90.476                  NA                     NA
#> 2   Skeletonema_marinoi   96.410                  NA                     NA
#> 3      Strombidium-like   92.808                  NA                     NA
#> 4                             NA                 495                 101547
#> 5                             NA                  NA                     NA
#>   UNCLASSIFIED_VOLUME  METOA                  COUNTPROG ALABO ACKR_ANA
#> 1                  NA IMA-SW ifcb-analysis, iRfcb 0.3.9  SMHI        N
#> 2                  NA IMA-SW ifcb-analysis, iRfcb 0.3.9  SMHI        N
#> 3                  NA IMA-SW ifcb-analysis, iRfcb 0.3.9  SMHI        N
#> 4           0.0272936 IMA-SW ifcb-analysis, iRfcb 0.3.9  SMHI        N
#> 5                  NA                                    SMHI        N
#>      ANADATE
#> 1 2024-09-11
#> 2 2024-09-11
#> 3 2024-09-11
#> 4 2024-09-11
#> 5 2022-07-13
#>                                                                                                                   METDC
#> 1 https://github.com/hsosik/ifcb-analysis, https://github.com/kudelalab/PSD, https://github.com/EuropeanIFCBGroup/iRfcb
#> 2 https://github.com/hsosik/ifcb-analysis, https://github.com/kudelalab/PSD, https://github.com/EuropeanIFCBGroup/iRfcb
#> 3 https://github.com/hsosik/ifcb-analysis, https://github.com/kudelalab/PSD, https://github.com/EuropeanIFCBGroup/iRfcb
#> 4 https://github.com/hsosik/ifcb-analysis, https://github.com/kudelalab/PSD, https://github.com/EuropeanIFCBGroup/iRfcb
#> 5                                                                                                                      
#>                                      TRAINING_SET TRAINING_SET_ANNOTATED_BY
#> 1 https://doi.org/10.17044/scilifelab.25883455.v3          Ann-Turi Skjevik
#> 2 https://doi.org/10.17044/scilifelab.25883455.v3          Ann-Turi Skjevik
#> 3 https://doi.org/10.17044/scilifelab.25883455.v3          Ann-Turi Skjevik
#> 4 https://doi.org/10.17044/scilifelab.25883455.v3          Ann-Turi Skjevik
#> 5                                                                          
#>   CLASSIFIER_CREATED_BY             CLASSIFIER_USED MANUAL_QC_DATE
#> 1    Anders Torstensson SMHI-RF-Baltic-v8-11Sep2024     2024-09-27
#> 2    Anders Torstensson SMHI-RF-Baltic-v8-11Sep2024     2024-09-27
#> 3    Anders Torstensson SMHI-RF-Baltic-v8-11Sep2024     2024-09-27
#> 4    Anders Torstensson SMHI-RF-Baltic-v8-11Sep2024     2024-09-27
#> 5                                                                 
#>   PRE_FILTER_SIZE PH_FB CHL_FB CDOM_FB PHYC_FB PHER_FB WATERFLOW_FB TURB_FB
#> 1             150    NA     NA      NA      NA      NA           NA      NA
#> 2             150    NA     NA      NA      NA      NA           NA      NA
#> 3             150    NA     NA      NA      NA      NA           NA      NA
#> 4             150    NA     NA      NA      NA      NA           NA      NA
#> 5              NA 7.623  0.556  3.0522      NA   6.685       9.7167  1.1529
#>   PCO2_FB TEMP_FB PSAL_FB OSAT_FB DOXY_FB
#> 1      NA      NA      NA      NA      NA
#> 2      NA      NA      NA      NA      NA
#> 3      NA      NA      NA      NA      NA
#> 4      NA      NA      NA      NA      NA
#> 5      NA 17.9333 10.8873 92.5014  5.7337