Zip PNG Folders

This function zips directories containing .png files and optionally includes README and MANIFEST files.

Usage

ifcb_zip_pngs(
  png_folder,
  zip_filename,
  readme_file = NULL,
  email_address = "",
  version = "",
  print_progress = TRUE
)

Arguments

png_folder

The directory containing subdirectories with .png files.

zip_filename

The name of the zip file to create.

readme_file

Optional path to a README file for inclusion in the zip package.

email_address

Optional email address to include in the README file.

version

Optional version information to include in the README file.

print_progress

A logical value indicating whether to print progress bar. Default is TRUE.

Value

This function does not return any value; it creates a zip archive and potentially a README file.

See also

Examples

if (FALSE) { # \dontrun{
# Zip all subdirectories in the 'images' folder with a README file
ifcb_zip_pngs("path/to/images",
             "images.zip",
             readme_file = system.file("exdata/README-template.md", package = "iRfcb"),
             email_address = "example@example.com",
             version = "1.0")

# Zip all subdirectories in the 'images' folder without a README file
ifcb_zip_pngs("path/to/images", "images.zip")
} # }